requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
python-decouple==3.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiosmtplib==3.0.1
aiosqlite==0.19.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
